move files to  folders for better organization
