+++ patientsearch/src/js/components/EditButtonGroup.js
@@ -9,7 +9,7 @@ import SaveIcon from "@material-ui/icons/Save";
 import Button from "@material-ui/core/Button";
 import ButtonGroup from "@material-ui/core/ButtonGroup";
 import Modal from "@material-ui/core/Modal";
-import theme from "../context/theme";
+import theme from "../themes/theme";
 
 const useStyles = makeStyles({
     buttonGroupContainer: {
