+++ patientsearch/src/js/components/Info.js
@@ -7,7 +7,7 @@ import CircularProgress from "@material-ui/core/CircularProgress";
 import Typography from "@material-ui/core/Typography";
 import { imageOK } from "./Utility";
 import { useSettingContext } from "../context/SettingContextProvider";
-import theme from "../context/theme";
+import theme from "../themes/theme";
 
 const useStyles = makeStyles({
   wrapper: {
