+++ patientsearch/src/js/components/SiteLogo.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { makeStyles } from "@material-ui/core/styles";
 import { imageOK } from "./Utility";
 import { useSettingContext } from "../context/SettingContextProvider";
-import theme from "../context/theme";
+import theme from "../themes/theme";
 
 const useStyles = makeStyles({
   container: {
