+++ patientsearch/src/js/components/TimeoutModal.js
@@ -3,7 +3,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import Button from "@material-ui/core/Button";
 import Modal from "@material-ui/core/Modal";
 import { sendRequest } from "./Utility";
-import theme from "../context/theme";
+import theme from "../themes/theme";
 import { useSettingContext } from "../context/SettingContextProvider";
 
 function getModalStyle() {
