+++ patientsearch/src/js/containers/Entry.js
