+++ patientsearch/src/js/layout/Layout.js
@@ -2,10 +2,10 @@ import PropTypes from "prop-types";
 import React from "react";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { ThemeProvider } from "@material-ui/styles";
-import Header from "./Header";
-import SystemBanner from "./SystemBanner";
+import Header from "../components/Header";
+import SystemBanner from "../components/SystemBanner";
 import SettingContextProvider from "../context/SettingContextProvider";
-import theme from "../context/theme";
+import theme from "../themes/theme";
 import "../../styles/app.scss";
 
 export default function Layout({children}) {
